#327 #335 Enabled 'property usage' logic in all tick managers.
